gui.py and data.json
The GUI now has basic functionality. It has label and entry widgets for every needed parameter, a default for each parameter, and has a button to save all of the variables and their values to a JSON file. Currently the dictionary saved has the parameter name as a String and the parameter value as a String. Included the JSON file the program created.