gui.py
Switched from keeping labels and entries in an array to a dictionary. Used .grid() after creating so the return values wouldn't be "None". Added tabs to switch between all parameters, part parameters, layer parameters, and file parameters. Added feature to input name of JSON file to save as or to upload.