windows and rearranging, start of gcode conversion
Save and Upload options now pull up a window to pick the directory to save in. Rearranged all of the buttons to make more sense. Deleted currPath, outputSubdirectory, startEndSubDirectory, and outputFileName. Added the button and started the function for converting the file to Gcode. Included .json file program now makes.